+++ src/networkNameDescriptor.cpp
@@ -8,7 +8,7 @@ bool NetworkNameDescriptor::unpack(Common::ReadStream& stream)
         return false;
     }
 
-    int size = stream.leftBytes();
+    size_t size = stream.leftBytes();
     networkName.resize(size);
     stream.read(networkName.data(), size);
     return true;
